labs/requirements.txt
Flask==2.1.1
Pillow==9.0.1
numpy==1.22.3
tensorflow==2.7.0
tensorflow-datasets==4.5.0
tensorflow-hub==0.12.0
flask-cors
gunicorn
scikit-learn
zipfile38
flask-sqlalchemy
flask-migrate
mysqlclient
Flask-Mail